Move extra json utils tests
Intent: consolidate json utils tests for clearer coverage
Operators involved: [Self-organization]
Affected invariants: [#8]
Key changes:
- Merge tests from test_json_utils_extra into test_json_utils
- Reuse DummyOrjson and _reset_json_utils fixtures
- Remove obsolete test_json_utils_extra file
Expected risks/dissonances: warning state handling; contained by targeted tests
Metrics: C(t), Si, νf, phase unchanged
Equivalence map: none
